Add /opt/homebrew to PATH if it exists
On Intel Macs, Homebrew installs into /usr/local which is, conveniently,
already on the PATH. On Apple Silicon Macs, however, Homebrew installs
into /opt/homebrew, which means it needs to be added to the PATH.

I’ve chucked in all the lines that the Homebrew installer said I should
add to .zprofile (the output of `brew shellenv`). And wrapped it all in
a check for the /opt/homebrew directory so this might, theoretically,
still work on Intel Macs too.